requirements_dev.txt
# Astrology Platform - Development Requirements (Windows Compatible)

# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# HTTP & WebSocket
requests==2.31.0
httpx==0.25.1
websockets==12.0

# Email
python-multipart==0.0.6

# Payment Gateways (Optional)
stripe==7.4.0

# Development Tools
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Production Server
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0
